requirements.txt
telethon>=1.34,<2.0
playwright>=1.40
pyyaml>=6.0
